Update DrillsIndustry to new design
