+++ src/main/java/zabbum/oelremake/Oilfield.java
@@ -9,6 +9,7 @@ import com.googlecode.lanterna.gui2.Panel;
 import com.googlecode.lanterna.gui2.table.Table;
 import lombok.Data;
 import zabbum.oelremake.Cars.CarsIndustry;
+import zabbum.oelremake.Drills.DrillsIndustry;
 
 import java.util.Random;
